+++ l10n_it_fatturapa_pec/models/fatturapa_attachment_out.py
@@ -10,7 +10,6 @@ from lxml import etree
 
 from odoo import api, fields, models, _
 from odoo.exceptions import UserError
-from odoo.addons.base.models.ir_mail_server import MailDeliveryException
 
 _logger = logging.getLogger(__name__)
 
@@ -25,68 +24,6 @@ class FatturaPAAttachmentOut(models.Model):
         string='Last Response from Exchange System', default='No response yet',
         readonly=True)
     sending_user = fields.Many2one("res.users", "Sending User", readonly=True)
-
-    @api.model
-    def _check_fetchmail(self):
-        server = self.env['fetchmail.server'].search([
-            ('is_fatturapa_pec', '=', True),
-            ('state', '=', 'done')
-        ])
-        if not server:
-            raise UserError(_(
-                "No incoming PEC server found. Please configure it."))
-
-    @api.multi
-    def send_via_pec(self):
-        self._check_fetchmail()
-        self.env.user.company_id.sdi_channel_id.check_first_pec_sending()
-        states = self.mapped('state')
-        if set(states) != set(['ready']):
-            raise UserError(
-                _("You can only send files in 'Ready to Send' state.")
-            )
-        for att in self:
-            if not att.datas or not att.datas_fname:
-                raise UserError(_("File content and file name are mandatory"))
-            mail_message = self.env['mail.message'].create({
-                'model': self._name,
-                'res_id': att.id,
-                'subject': att.name,
-                'body': 'XML file for FatturaPA {} sent to Exchange System to '
-                        'the email address {}.'
-                .format(
-                    att.name,
-                    self.env.user.company_id.email_exchange_system),
-                'attachment_ids': [(6, 0, att.ir_attachment_id.ids)],
-                'email_from': (
-                    self.env.user.company_id.email_from_for_fatturaPA),
-                'reply_to': (
-                    self.env.user.company_id.email_from_for_fatturaPA),
-                'mail_server_id': self.env.user.company_id.sdi_channel_id.
-                pec_server_id.id,
-            })
-
-            mail = self.env['mail.mail'].create({
-                'mail_message_id': mail_message.id,
-                'body_html': mail_message.body,
-                'email_to': self.env.user.company_id.email_exchange_system,
-                'headers': {
-                    'Return-Path':
-                    self.env.user.company_id.email_from_for_fatturaPA
-                }
-            })
-
-            if mail:
-                try:
-                    mail.send(raise_exception=True)
-                    att.state = 'sent'
-                    att.sending_date = fields.Datetime.now()
-                    att.sending_user = self.env.user.id
-                    self.env.user.company_id.sdi_channel_id.\
-                        update_after_first_pec_sending()
-                except MailDeliveryException as e:
-                    att.state = 'sender_error'
-                    mail.body = str(e)
 
     @api.multi
     def parse_pec_response(self, message_dict):
