+++ l10n_it_fatturapa_pec/tests/test_e_invoice_response.py
@@ -29,7 +29,7 @@ class TestEInvoiceResponse(EInvoiceCommon):
         e-invoice to 'validated'"""
         e_invoice = self._create_e_invoice()
         self.set_e_invoice_file_id(e_invoice, 'IT03339130126_00009.xml')
-        e_invoice.send_via_pec()
+        e_invoice.send_to_sdi()
 
         incoming_mail = self._get_file(
             'POSTA CERTIFICATA_ Ricevuta di consegna 6782414.txt')
@@ -43,7 +43,7 @@ class TestEInvoiceResponse(EInvoiceCommon):
         """Receiving a 'CONSEGNA' posts a mail.message in the e-invoice"""
         e_invoice = self._create_e_invoice()
         self.set_e_invoice_file_id(e_invoice, 'IT03339130126_00009.xml')
-        e_invoice.send_via_pec()
+        e_invoice.send_to_sdi()
 
         incoming_mail = self._get_file(
             'CONSEGNA_ IT03339130126_00009.xml.txt')
@@ -66,7 +66,7 @@ class TestEInvoiceResponse(EInvoiceCommon):
         """Receiving a 'ACCETTAZIONE' posts a mail.message in the e-invoice"""
         e_invoice = self._create_e_invoice()
         self.set_e_invoice_file_id(e_invoice, 'IT03339130126_00009.xml')
-        e_invoice.send_via_pec()
+        e_invoice.send_to_sdi()
 
         incoming_mail = self._get_file(
             'ACCETTAZIONE_ IT03339130126_00009.xml.txt')
@@ -145,7 +145,7 @@ class TestEInvoiceResponse(EInvoiceCommon):
         self.set_sequences(2621, '2019-01-08')
         e_invoice = self._create_e_invoice()
         self.set_e_invoice_file_id(e_invoice, 'IT14627831002_02621.xml')
-        e_invoice.send_via_pec()
+        e_invoice.send_to_sdi()
 
         incoming_mail = self._get_file(
             'POSTA CERTIFICATA_mancata_consegna.txt')
