+++ l10n_it_fatturapa_pec/views/fatturapa_attachment_out.xml
@@ -6,10 +6,6 @@
         <field name="model">fatturapa.attachment.out</field>
         <field name="inherit_id" ref="l10n_it_fatturapa_out.view_fatturapa_out_attachment_form"/>
         <field name="arch" type="xml">
-            <xpath expr="//header" position="inside">
-                    <button name="send_via_pec" states="ready" string="Send Via PEC" type="object" class="oe_highlight"/>
-                    <button name="reset_to_ready" states="sender_error" string="Reset to ready" type="object" />
-            </xpath>
             <xpath expr="//div/group[last()]" position="after">
                 <group name="last_sdi_response" string="Last Response from Exchange System" attrs="{'invisible':[('state', '=', 'ready')]}">
                     <field name="last_sdi_response" nolabel="1"/>
