--- l10n_it_fatturapa_pec/wizard/__init__.py
@@ -1,2 +1,0 @@
-
-from . import send_pec
